requirements.txt
openai
PyQt5
python-dotenv
